Prevent users from manually hitting routes they do not have access to

+++ routes/galleryRoutes.js
@@ -3,43 +3,38 @@ const express = require('express');
 const router = express.Router();
 const db = require('../models');
 
+function isOwner(req, data) {
+  return req.isAuthenticated() && req.user.dataValues.id === data.dataValues.UserId;
+}
 
 router.route('/')
       // Post a new image to the gallery with relevant info and associate the user with the image.
       // After success, redirect to gallery.
       .post((req, res) => {
-        console.log(req.user.dataValues);
-        db.Picture.create({
-          author: req.body.author,
-          link: req.body.link,
-          description: req.body.description,
-          UserId: parseInt(req.user.dataValues.id)
-        }).then(function(){
-          res.redirect('/');
-        });
+        if (req.isAuthenticated()) {
+          db.Picture.create({
+            author: req.body.author,
+            link: req.body.link,
+            description: req.body.description,
+            UserId: parseInt(req.user.dataValues.id)
+          })
+          .then(function(){
+            res.redirect('/');
+          });
+        }
       })
       // Retrieves the index page
       // If the user is authenticated, show the edit and delete buttons for pictures.
       .get((req, res) => {
         db.Picture.findAll()
         .then((data) => {
-
-
-
-          if(req.isAuthenticated()=== true){
-
-            let isOwnerArr = data.map((v) => {
-              return req.user.dataValues.id === v.dataValues.UserId;
-            });
-
-            console.log(isOwnerArr);
-
+          if (req.isAuthenticated()) {
             res.render('index', {
               picture: data,
               loggedin: true
             });
           } else {
-              res.render('index', {
+            res.render('index', {
               picture: data
             });
           }
@@ -62,9 +57,9 @@ router.route('/:id')
           })
           // If a user is authenticated, add edit and delete buttons on the picture page.
           .then((data) => {
-            if(req.isAuthenticated()=== true){
+            if (req.isAuthenticated()) {
               data.dataValues.loggedIn = true;
-              data.dataValues.isOwner = (req.user.dataValues.id === data.dataValues.UserId);
+              data.dataValues.isOwner = isOwner(req, data);
                 res.render('picture', data.dataValues);
             } else {
             res.render('picture', data.dataValues);
@@ -73,31 +68,48 @@ router.route('/:id')
         })
         // Update an image
         .put((req, res) => {
-          db.Picture.update({
-            author: req.body.author,
-            link: req.body.link,
-            description: req.body.description
-          }, {
+          db.Picture.findOne({
             where: {
               id: req.params.id
             }
           })
-          .then((data)=>{
-            res.redirect(303,`/gallery/${req.params.id}`);
+          .then((data) => {
+            if (isOwner(req, data)) {
+              db.Picture.update({
+                author: req.body.author,
+                link: req.body.link,
+                description: req.body.description
+              }, {
+                where: {
+                  id: req.params.id
+                }
+              })
+              .then((data)=>{
+                res.redirect(303,`/gallery/${req.params.id}`);
+              });
+            }
           });
         })
         // Delete an image
         .delete( (req, res) => {
-          db.Picture.destroy({
+          db.Picture.findOne({
             where: {
               id: req.params.id
             }
           })
-          .then( (data) => {
-            res.redirect(303, '/gallery');
+          .then((data) => {
+            if (isOwner(req, data)) {
+              db.Picture.destroy({
+                where: {
+                  id: req.params.id
+                }
+              })
+              .then( (data) => {
+                res.redirect(303, '/gallery');
+              });
+            }
           });
         });
-
 
 router.route('/:id/edit')
         // Render picture editing page.
@@ -108,7 +120,7 @@ router.route('/:id/edit')
             }
           })
           .then((data) => {
-            if(req.isAuthenticated()=== true){
+            if (isOwner(req, data)) {
               res.render('edit', {
                 id: req.params.id,
                 author: data.dataValues.author,
